requirements.txt
# =============================================================================
# Python Dependencies - E-Commerce Intelligence Platform
# =============================================================================

# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# Data Processing
pyspark==3.5.0
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.2
delta-spark==3.0.0

# ML & Model Serving
scikit-learn==1.3.2
mlflow==2.10.0
bentoml==1.1.11

# MinIO / S3
boto3==1.34.0
minio==7.2.3

# Monitoring
prometheus-client==0.19.0

# Data Generation
faker==22.0.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
click==8.1.7

# Testing
pytest==7.4.4
locust==2.20.0

# Kaggle CLI
kaggle==1.6.3
